+++ packages/tabs/src/Tab/index.ts
@@ -1,0 +1,3 @@
+export { default as InternalTab } from './InternalTab';
+export { default } from './Tab';
+export type { TabProps } from './Tab.types';
