+++ packages/tabs/src/Tabs/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Tabs';
+export type { AccessibleTabsProps, TabsProps } from './Tabs.types';
