+++ packages/tabs/src/index.ts
@@ -1,3 +1,4 @@
+export type { TabProps } from './Tab';
 export { default as Tab } from './Tab';
+export type { AccessibleTabsProps as TabsProps } from './Tabs';
 export { default as Tabs } from './Tabs';
-export type { TabProps, AccessibleTabsProps as TabsProps } from './types';
